move DirectWriteResources to UI
