+++ WeaselUI/Layout.cpp
@@ -1,5 +1,6 @@
 #include "stdafx.h"
 #include "Layout.h"
+using namespace weasel;
 
 Layout::Layout(const UIStyle& style, const Context& context, const Status& status)
 	: _style(style), _context(context), _status(status),
