+++ WeaselUI/Layout.h
@@ -3,7 +3,6 @@
 #include <WeaselCommon.h>
 #include <WeaselUI.h>
 #include <gdiplus.h>
-#include "fontClasses.h"
 
 #pragma comment(lib, "gdiplus.lib")
 #define IS_FULLSCREENLAYOUT(style)	(style.layout_type == UIStyle::LAYOUT_VERTICAL_FULLSCREEN || style.layout_type == UIStyle::LAYOUT_HORIZONTAL_FULLSCREEN)
@@ -52,7 +51,7 @@ namespace weasel
 	public:
 		Layout(const UIStyle &style, const Context &context, const Status &status);
 
-		virtual void DoLayout(CDCHandle dc, DirectWriteResources* pDWR = NULL) = 0;
+		virtual void DoLayout(CDCHandle dc, PDWR pDWR = NULL) = 0;
 		/* All points in this class is based on the content area */
 		/* The top-left corner of the content area is always (0, 0) */
 		virtual CSize GetContentSize() const = 0;
@@ -73,7 +72,7 @@ namespace weasel
 		virtual std::wstring GetLabelText(const std::vector<Text> &labels, int id, const wchar_t *format) const = 0;
 		virtual bool IsInlinePreedit() const = 0;
 		virtual bool ShouldDisplayStatusIcon() const = 0;
-		virtual void GetTextSizeDW(const std::wstring text, size_t nCount, IDWriteTextFormat1* pTextFormat, DirectWriteResources* pDWR,  LPSIZE lpSize) const = 0;
+		virtual void GetTextSizeDW(const std::wstring text, size_t nCount, IDWriteTextFormat1* pTextFormat, PDWR pDWR,  LPSIZE lpSize) const = 0;
 		
 		int offsetX = 0;
 		int offsetY = 0;
