+++ include/WeaselUI.h
@@ -1,9 +1,24 @@
 ﻿#pragma once
 
 #include <WeaselCommon.h>
+#include <vector>
+#include <regex>
+#include <iterator>
+#include <d2d1.h>
+#include <dwrite_2.h>
+#include <memory>
 
 namespace weasel
 {
+
+	template <class T> void SafeRelease(T** ppT)
+	{
+		if (*ppT)
+		{
+			(*ppT)->Release();
+			*ppT = NULL;
+		}
+	}
 
 	enum ClientCapabilities
 	{
@@ -11,7 +26,11 @@ namespace weasel
 	};
 
 	class UIImpl;
+	class DirectWriteResources;
+	template <class T>
+	using an = std::shared_ptr<T>;
 
+	using PDWR = an<DirectWriteResources>;
 	//
 	// 输入法界面接口类
 	//
@@ -25,6 +44,8 @@ namespace weasel
 		{
 			if (pimpl_)
 				DestroyAll();
+			if (pDWR)
+				pDWR.reset();
 		}
 
 		// 创建输入法界面
@@ -56,9 +77,11 @@ namespace weasel
 		Status& status() { return status_; } 
 		UIStyle& style() { return style_; }
 		UIStyle& ostyle() { return ostyle_; }
+		PDWR pdwr() { return pDWR; }
 
 	private:
 		UIImpl* pimpl_;
+		PDWR pDWR;
 
 		Context ctx_;
 		Context octx_;
@@ -67,4 +90,29 @@ namespace weasel
 		UIStyle ostyle_;
 	};
 
+	class DirectWriteResources
+	{
+	public:
+		DirectWriteResources(weasel::UIStyle& style, UINT dpi);
+		~DirectWriteResources();
+
+		HRESULT InitResources(std::wstring label_font_face, int label_font_point,
+			std::wstring font_face, int font_point,
+			std::wstring comment_font_face, int comment_font_point, bool vertical_text = false);
+		HRESULT InitResources(UIStyle& style, UINT dpi);
+		void SetDpi(UINT dpi);
+		float dpiScaleX_, dpiScaleY_;
+		ID2D1Factory* pD2d1Factory;
+		IDWriteFactory2* pDWFactory;
+		ID2D1DCRenderTarget* pRenderTarget;
+		IDWriteTextFormat1* pPreeditTextFormat;
+		IDWriteTextFormat1* pTextFormat;
+		IDWriteTextFormat1* pLabelTextFormat;
+		IDWriteTextFormat1* pCommentTextFormat;
+		IDWriteTextLayout2* pTextLayout;
+	private:
+		UIStyle& _style;
+		void _ParseFontFace(const std::wstring fontFaceStr, DWRITE_FONT_WEIGHT& fontWeight, DWRITE_FONT_STYLE& fontStyle);
+		void _SetFontFallback(IDWriteTextFormat1* pTextFormat, std::vector<std::wstring> fontVector);
+	};
 }
